refactor: flatten message processor module
