+++ src/index.ts
@@ -6,7 +6,7 @@ import { deployGlobalCommands } from './commands/deploy-command';
 import { getConfigs } from './config';
 import { loadEnv } from './utils/load-env';
 import { logger } from './utils/logger';
-import { processMessage } from './utils/messageProcessor';
+import { processMessage } from './utils/message-processor';
 
 const main = async () => {
   loadEnv();
