+++ src/utils/message-processor.test.ts
@@ -1,6 +1,6 @@
 import type { Message } from 'discord.js';
 import { describe, expect, it, vi } from 'vitest';
-import { type CommandConfig, processMessage } from '.';
+import { type CommandConfig, processMessage } from './message-processor';
 
 describe('processMessage', () => {
   beforeEach(() => {
